convert more 'unsigned int' to 'size_t' and 'int' to 'ssize_t'

+++ src/driver.h
@@ -93,8 +93,8 @@ typedef float _aaxDriverParam(const void*, enum _aaxDriverParam);
 typedef char *_aaxDriverGetDevices(const void*, int mode);
 typedef char *_aaxDriverGetInterfaces(const void*, const char*, int mode);
 
-typedef int _aaxDriverCallback(const void*, void*, float, float);
-typedef int _aaxDriverCaptureCallback(const void*, void**, int*, size_t*, void*, size_t, float);
+typedef size_t _aaxDriverCallback(const void*, void*, float, float);
+typedef size_t _aaxDriverCaptureCallback(const void*, void**, ssize_t*, size_t*, void*, size_t, float);
 
 typedef void _aaxDriverPrepare3d(void*, const void*, float, float, void*, void*);
 typedef void _aaxDriverPostProcess(const void*, void*, const void*);
